Extracted _create_all_compiler_instances_targets_build_dirs func
